+++ ChangeLog.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - thread type support to the function `thread` to allow spawning daemon or user 
   threads
 - support for tracing exceptions in `trace/trace-var`
+- math function `rand-bigint`